+++ packages/core/src/utils/sourceMaps.test.ts
@@ -1,0 +1,36 @@
+import { getBundleId, getBundleIdFromError, getBundleIdStackMap } from './sourceMaps';
+
+describe('sourceMapUpload utils', () => {
+  beforeEach(() => {
+    delete (global as any).__faroBundleId_foo;
+    delete (global as any).__faroBundleIds;
+  });
+
+  afterAll(() => {
+    delete (global as any).__faroBundleId_foo;
+    delete (global as any).__faroBundleIds;
+  });
+
+  it('can get the bundle ID from the global object', () => {
+    expect(getBundleId('foo')).toBeUndefined();
+
+    (global as any).__faroBundleId_foo = 'bar';
+    expect(getBundleId('foo')).toEqual('bar');
+  });
+
+  it('can get the bundle ID stack map from the global object', () => {
+    expect(getBundleIdStackMap()).toBeUndefined();
+
+    const e = new Error();
+    (global as any).__faroBundleIds = new Map([[e, 'bar']]);
+    expect(getBundleIdStackMap()).toEqual(new Map([[e, 'bar']]));
+  });
+
+  it('can get the bundle ID from an error', () => {
+    const e = new Error();
+    expect(getBundleIdStackMap()).toBeUndefined();
+
+    (global as any).__faroBundleIds = new Map([[e, 'bar']]);
+    expect(getBundleIdFromError(e)).toEqual('bar');
+  });
+});