+++ packages/web-sdk/src/instrumentations/errors/stackFrames/buildStackFrame.ts
@@ -6,7 +6,8 @@ export function buildStackFrame(
   filename: string | undefined,
   func: string | undefined,
   lineno: number | undefined,
-  colno: number | undefined
+  colno: number | undefined,
+  bundleid?: string | undefined
 ): ExceptionStackFrame {
   const stackFrame: ExceptionStackFrame = {
     filename: filename || document.location.href,
@@ -19,6 +20,10 @@ export function buildStackFrame(
 
   if (colno !== undefined) {
     stackFrame.colno = colno;
+  }
+
+  if (bundleid !== undefined) {
+    stackFrame.bundleid = bundleid;
   }
 
   return stackFrame;